v 1.3

+++ src/main/java/com/ff/filip/jsf/mb/MBIspit.java
@@ -7,13 +7,25 @@ package com.ff.filip.jsf.mb;
 
 import com.ff.filip.domen.Ispit;
 import com.ff.filip.jpa.dbb.IspitService;
+import java.io.IOException;
 import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import javax.enterprise.context.SessionScoped;
 import javax.faces.application.FacesMessage;
 import javax.faces.context.FacesContext;
 import javax.inject.Inject;
 import javax.inject.Named;
+import javax.servlet.ServletOutputStream;
+import javax.servlet.http.HttpServletResponse;
+import net.sf.jasperreports.engine.JRException;
+import net.sf.jasperreports.engine.JasperExportManager;
+import net.sf.jasperreports.engine.JasperFillManager;
+import net.sf.jasperreports.engine.JasperPrint;
+import net.sf.jasperreports.engine.data.JRBeanCollectionDataSource;
 
 /**
  *
@@ -30,6 +42,7 @@ public class MBIspit implements Serializable {
     private Ispit ispitForEditing;
     private Ispit ispitForDeleting;
     private List<Ispit> listIspit;
+    private boolean isInitialized = false;
 
     public MBIspit() {
         ispit = new Ispit();
@@ -59,6 +72,7 @@ public class MBIspit implements Serializable {
 
     public void setIspitForDeleting(Ispit ispitForDeleting) {
         this.ispitForDeleting = ispitForDeleting;
+        isInitialized = true;
         FacesContext.getCurrentInstance().addMessage(null, new FacesMessage(FacesMessage.SEVERITY_WARN, "Upozorenje! ", "  Ispit je postavljen za brisanje."));
     }
 
@@ -75,8 +89,12 @@ public class MBIspit implements Serializable {
     }
 
     public void deleteIspit() {
-        ispitService.deleteIspitById(ispitForDeleting);
-        FacesContext.getCurrentInstance().addMessage(null, new FacesMessage(FacesMessage.SEVERITY_INFO, "INFO: ", "  Ispit je uspesno obrisan."));
+        if (isInitialized) {
+            ispitService.deleteIspitById(ispitForDeleting);
+            FacesContext.getCurrentInstance().addMessage(null, new FacesMessage(FacesMessage.SEVERITY_INFO, "INFO: ", "  Ispit je uspesno obrisan."));
+        } else {
+            FacesContext.getCurrentInstance().addMessage(null, new FacesMessage(FacesMessage.SEVERITY_WARN, "Upozorenje!: ", "  Niste postavili ispit za brisanje"));
+        }
     }
 
     public void persistIspit() {
@@ -103,6 +121,31 @@ public class MBIspit implements Serializable {
         flag = ispitService.checkIspitPass(ispit);
 
         return flag;
+    }
+
+    public void printPDFIspit() throws JRException, IOException {
+        Map<String, Object> parameters = new HashMap<String, Object>();
+        Date date = new Date();
+        System.out.println("usao2");
+        List<Ispit> forPrinting = new ArrayList<>();
+        findAllIspit();
+        forPrinting = this.listIspit;
+        for (Ispit i : forPrinting) {
+            System.out.println(i.getSifraIspita());
+        }
+
+        JRBeanCollectionDataSource beanCollectionDataSource = new JRBeanCollectionDataSource(forPrinting);
+
+        String reportPath = FacesContext.getCurrentInstance().getExternalContext().getRealPath("/reports/IspitMaster.jasper");
+        JasperPrint jasperPrint = JasperFillManager.fillReport(reportPath, parameters, beanCollectionDataSource);
+
+        HttpServletResponse httpServletResponse = (HttpServletResponse) FacesContext.getCurrentInstance().getExternalContext().getResponse();
+        httpServletResponse.setContentType("application / pdf");
+        httpServletResponse.addHeader("Content-disposition", "attachment; filename=StudentMaster_" + date + ".pdf");
+        ServletOutputStream servletOutputStream = httpServletResponse.getOutputStream();
+
+        JasperExportManager.exportReportToPdfStream(jasperPrint, servletOutputStream);
+        FacesContext.getCurrentInstance().responseComplete();
     }
 
 }
